[profiler] Split CUPTI monitor tests into test_cupti_monitor.py
