+++ test/run_test.py
@@ -219,6 +219,7 @@ S390X_BLOCKLIST = [
     "inductor/test_torchinductor_codegen_dynamic_shapes",
     "lazy/test_meta_kernel",
     "onnx/test_utility_funs",
+    "profiler/test_cupti_monitor",
     "profiler/test_profiler",
     "test_jit",
     "dynamo/test_utils",
@@ -1824,6 +1825,7 @@ def get_selected_tests(options) -> list[str]:
                     "functorch/test_aotdispatch",
                     "functorch/test_control_flow",
                     "nn/test_convolution",
+                    "profiler/test_cupti_monitor",
                     "profiler/test_profiler",
                     "test_modules",
                     "test_expanded_weights",
